+++ quibble/commands.py
@@ -13,7 +13,6 @@ import yaml
 
 from quibble.gitchangedinhead import GitChangedInHead
 from quibble.util import copylog, isExtOrSkin, ProgressReporter
-from quibble.phpunit_split import Splitter
 import quibble.mediawiki.registry
 import quibble.zuul
 import subprocess
@@ -990,109 +989,6 @@ class PhpUnitPrepareParallelRunComposer:
         return "PHPUnit Prepare Parallel Run (Composer)"
 
 
-class PhpUnitPrepareParallelRun:
-    """To run tests in parallel, we need to split the tests that
-    will be run into smaller suites that we can execute individually
-    and in parallel. This command runs the phpunit `--list-tests-xml`
-    function, which dumps out a list of which test classes would be
-    included in a run of the provided test suite.
-
-    Note that this command is not sensitive to `--group` or
-    `--exclude-group` - that filtering happens later when the suites
-    are actually executed. This can lead to the situation of a test
-    class being included where no tests are then run.
-
-    We dump the list of tests into the file
-    `test-cases-integration.xml`, which is then picked up by
-    `phpunit_split.Splitter` to create the new, smaller suites that
-    we will use. `Splitter` copies `phpunit.dist.xml` to
-    `phpunit.xml` and adds new test suites named `split_group_X`
-    where `X` is a number.
-
-    Once the modified `phpunit.xml` exists, it has priority over
-    `phpunit.dist.xml` and is used by default by `phpunit`
-    commands. This also means that we can run the split test suites
-    normally from the command line with the `--testsuite` argument
-    to `phpunit`.
-
-    @see T365978
-    @deprecated - this implementation is deprecated since the
-    introduction of `PhpUnitPrepareParallelRunComposer`, and will
-    be removed in a future release.
-    """
-
-    def __init__(
-        self,
-        mw_install_path,
-        testsuite='extensions',
-        log_dir=None,
-        junit=False,
-    ):
-        self.mw_install_path = mw_install_path
-        self.testsuite = testsuite
-        self.log_dir = log_dir
-        self.junit = junit
-
-    def _do_generate_test_list(self):
-        """Use phpunit's `--list-tests-xml` function to generate a
-        list of test classes that would be included in the suite"""
-        phpunit_env = {}
-        phpunit_env.update(os.environ)
-        phpunit_env.update({'LANG': 'C.UTF-8'})
-
-        phpunit_command = [
-            'composer',
-            'run',
-            '--timeout=0',
-            'phpunit:entrypoint',
-            '--',
-        ]
-
-        if self.testsuite is not None:
-            phpunit_command.append('--testsuite')
-            phpunit_command.append(self.testsuite)
-
-        phpunit_command.append('--list-tests-xml')
-        phpunit_command.append('test-cases-integration.xml')
-
-        run(phpunit_command, cwd=self.mw_install_path, env=phpunit_env)
-
-    def _do_generate_split_suite(self):
-        """Split the test suite into smaller chunks.
-
-        We create 8 chunks - 7 chunks of split out tests, and 1 chunk
-        for the `ParserIntegrationTest.php` and `SandboxTest.php` tests
-        (which are their own special test generating magic). Anecdotally,
-        8 chunks is enough to max out the CPU on my development machine
-        (8 cores on an 11th Gen Intel(R) Core(TM) i7-11), and to arrange
-        that the chunks of split tests take approximately as long as the
-        `ParserIntegrationTest.php` suite of tests.
-
-        The result of the splitting will be the updated `phpunit.xml` file"""
-        splitter = Splitter(self.mw_install_path)
-        # Pass 7 here for 7 chunks. We pass None for the
-        # `phpunit.results.cache` file, since we're not using that (yet)
-        splitter.split(
-            os.path.join(self.mw_install_path, 'test-cases-integration.xml'),
-            7,
-            None,
-        )
-        # To support developers in reproducing failed test runs, we
-        # make a copy of the phpunit.xml file in the logs folder -
-        # this adds the file to the artefacts collected by Jenkins
-        copylog(
-            os.path.join(self.mw_install_path, 'phpunit.xml'),
-            os.path.join(self.log_dir, 'phpunit-parallel.xml'),
-        )
-
-    def execute(self):
-        self._do_generate_test_list()
-        self._do_generate_split_suite()
-
-    def __str__(self):
-        return "PHPUnit Prepare Parallel Run"
-
-
 class AbstractParallelPhpUnit:
     """Parent class for running the parallel phpunit test suites."""
 
@@ -1131,45 +1027,6 @@ class PhpUnitDatabaselessParallelComposer(AbstractParallelPhpUnit):
         ).format(self.testsuite or 'default')
 
 
-class PhpUnitDatabaselessParallel(AbstractParallelPhpUnit):
-    """Run the tests in the provided suite in parallel, excluding
-    Database and Standalone tests.
-
-    Will create 7 parallel runners, and requires
-    PhpUnitPrepareParallelRun to have been executed beforehand to
-    setup the `phpunit.xml` file with the expect `split_group_X`
-    suites.
-
-    Not that the splitting process places ExtensionsParserTestSuite
-    alone in split_group_7. Since all of the ExtensionsParserTestSuite
-    requires the database, we don't need to run that group here
-    (and doing so would cause PHPUnit to exit with a positive exit
-    status, failing the test run).
-    """
-
-    def _get_phpunit_command(self, split_group_id):
-        return PhpUnitDatabaseless(
-            self.mw_install_path,
-            'split_group_%s' % split_group_id,
-            self.log_dir,
-            self.junit,
-            os.path.join(
-                self.log_dir,
-                '.phpunit_group_%s_databaseless.cache.json' % split_group_id,
-            ),
-        )
-
-    def generate_parallel_command(self):
-        commands = [self._get_phpunit_command(i) for i in range(0, 7)]
-        return Parallel(name=str(self), steps=commands)
-
-    def __str__(self):
-        return (
-            "PHPUnit {} suite (without database "
-            "or standalone) parallel run".format(self.testsuite or 'default')
-        )
-
-
 class PhpUnitDatabaseParallelComposer(AbstractParallelPhpUnit):
     """Run the tests in the provided suite in parallel, excluding
     Standalone tests and including the Database tests."""
@@ -1194,42 +1051,6 @@ class PhpUnitDatabaseParallelComposer(AbstractParallelPhpUnit):
         return (
             "PHPUnit {} suite (with database) parallel run (Composer)"
         ).format(self.testsuite or 'default')
-
-
-class PhpUnitDatabaseParallel(AbstractParallelPhpUnit):
-    """Run the tests in the provided suite in parallel, excluding
-    Standalone tests but including Database tests.
-
-    Will create 8 parallel runners, and requires
-    PhpUnitPrepareParallelRun to have been executed beforehand to
-    setup the `phpunit.xml` file with the expect `split_group_X`
-    suites.
-
-    @deprecated - this implementation is deprecated since the
-    introduction of `PhpUnitDatabaseParallelComposer`, and will
-    be removed in a future release.
-    """
-
-    def _get_phpunit_command(self, split_group_id):
-        return PhpUnitDatabase(
-            self.mw_install_path,
-            'split_group_%s' % split_group_id,
-            self.log_dir,
-            self.junit,
-            os.path.join(
-                self.log_dir,
-                '.phpunit_group_%s_database.cache.json' % split_group_id,
-            ),
-        )
-
-    def generate_parallel_command(self):
-        commands = [self._get_phpunit_command(i) for i in range(0, 8)]
-        return Parallel(name=str(self), steps=commands)
-
-    def __str__(self):
-        return "PHPUnit {} suite (with database) parallel run".format(
-            self.testsuite or 'default'
-        )
 
 
 class PhpUnitParallelNotice: